smart-scaling-project/requirements.txt
flask
flask-cors
openstacksdk
python-keystoneclient
python-novaclient
python-heatclient
gnocchiclient